feat: Add JsonElementExtensions for converting JsonElement to Dictionary
feat: Introduce NumberExtensions for converting short and int to decimal percentages

feat: Enhance StringExtensions with methods for converting strings to Guid and nullable Guid

refactor: Update IHttpClientsService to use ApiResponse instead of IApiResponse

refactor: Modify CrudController to return ApiResponse instead of ApiEmptyResponse

refactor: Update DomainCrudController to return ApiResponse for CRUD operations

refactor: Change DomainTranslationController to return ApiResponse for save and delete actions

delete: Remove unused DecimalExtensions and GrpcResponseExtensions

delete: Remove IApiResponseExtensions and IEnumerableExtensions as they are no longer needed

delete: Remove StringExtensions from Specialized.Common due to duplication

fix: Update JsonConverter to handle null values correctly for short, byte, sbyte, and ushort

refactor: Update BaseCommand and BaseCommandHandler to use ApiResponse

refactor: Modify BaseQuery and BaseQueryHandler to use ApiResponse

refactor: Update FileHelper to use ApiResponse

refactor: Change BaseRepository methods to return ApiResponse instead of IApiResponse

refactor: Update PaginatedList extension methods to use ApiResponse

feat: Add EnumerableExtensions for translation-related methods

refactor: Update CrudService methods to return ApiResponse instead of ApiEmptyResponse

fix: Correct parameter handling in CrudService for AddParams method

refactor: Update ICrudService to return ApiResponse for CRUD operations

refactor: Change HttpClientService to return ApiResponse instead of IApiResponse

refactor: Update ITranslationService to return ApiResponse for message handling and domain translation operations

refactor: Modify TranslationService to return ApiResponse and streamline message handling

+++ src/BiUM.Specialized/Common/Mapper/JsonConverter.cs
@@ -116,10 +116,10 @@ public sealed class JsonShortConverter : JsonConverter<short>
 {
     public override short Read(ref Utf8JsonReader r, Type t, JsonSerializerOptions o)
     {
-        if (r.TokenType == JsonTokenType.Null) return (short)0;
+        if (r.TokenType == JsonTokenType.Null) return 0;
         if (JsonReadHelpers.TryReadInt64(ref r, out var v)) return (short)v;
 
-        return (short)0;
+        return 0;
     }
 
     public override void Write(Utf8JsonWriter w, short v, JsonSerializerOptions o) => w.WriteNumberValue(v);
@@ -129,10 +129,10 @@ public sealed class JsonByteConverter : JsonConverter<byte>
 {
     public override byte Read(ref Utf8JsonReader r, Type t, JsonSerializerOptions o)
     {
-        if (r.TokenType == JsonTokenType.Null) return (byte)0;
+        if (r.TokenType == JsonTokenType.Null) return 0;
         if (JsonReadHelpers.TryReadInt64(ref r, out var v)) return (byte)v;
 
-        return (byte)0;
+        return 0;
     }
 
     public override void Write(Utf8JsonWriter w, byte v, JsonSerializerOptions o) => w.WriteNumberValue(v);
@@ -142,10 +142,10 @@ public sealed class JsonSByteConverter : JsonConverter<sbyte>
 {
     public override sbyte Read(ref Utf8JsonReader r, Type t, JsonSerializerOptions o)
     {
-        if (r.TokenType == JsonTokenType.Null) return (sbyte)0;
+        if (r.TokenType == JsonTokenType.Null) return 0;
         if (JsonReadHelpers.TryReadInt64(ref r, out var v)) return (sbyte)v;
 
-        return (sbyte)0;
+        return 0;
     }
 
     public override void Write(Utf8JsonWriter w, sbyte v, JsonSerializerOptions o) => w.WriteNumberValue(v);
@@ -181,10 +181,10 @@ public sealed class JsonUShortConverter : JsonConverter<ushort>
 {
     public override ushort Read(ref Utf8JsonReader r, Type t, JsonSerializerOptions o)
     {
-        if (r.TokenType == JsonTokenType.Null) return (ushort)0;
+        if (r.TokenType == JsonTokenType.Null) return 0;
         if (JsonReadHelpers.TryReadInt64(ref r, out var v)) return (ushort)Math.Max(v, 0);
 
-        return (ushort)0;
+        return 0;
     }
 
     public override void Write(Utf8JsonWriter w, ushort v, JsonSerializerOptions o) => w.WriteNumberValue(v);
